set resource.contentType so that bundling works
Without setting contentType it was unable to determine it was a css
file and would not bundle it. Also added a functional test to ensure
that css file is bundled with less file in one file that matches the
name of the defaultBundle. Unfortunately the less file still ignores
defaultBundle, but at least we are getting closer.